refactor(codegen): extract omitted-param padding to a sibling module
Keeps extern_func.rs under the 2000-line CI file-size gate. The padding
logic added for #5812 item 4 moves verbatim into
lower_call/omitted_native_params.rs; the four manifest-param lowering
helpers it calls are bumped to pub(super).
