+++ crates/perry-codegen/src/lower_call/mod.rs
@@ -50,6 +50,7 @@ mod native_module_dispatch;
 mod native_table;
 mod new;
 mod new_helpers;
+mod omitted_native_params;
 mod options;
 mod property_get;
 mod ui_styling;
